+++ examples/exp_agedb_explore_data.ini
@@ -9,7 +9,7 @@ save = True
 databases = ['emodb']
 emodb = ./data/emodb/emodb
 emodb.split_strategy = random
-emodb.limit_samples = 20
+emodb.files_tables = ['emotion']
 target = age
 labels = ['u29', 'ü29']
 bins  = [-100000, 29, 100000]
