+++ examples/exp_emodb_audmodel_mlp.ini
@@ -8,9 +8,9 @@ save = True
 databases = ['emodb']
 emodb = ./data/emodb/emodb
 emodb.split_strategy = speaker_split
+emodb.files_tables = ['emotion']
 emodb.mapping = {'anger':'angry', 'happiness':'happy', 'sadness':'sad', 'neutral':'neutral'}
 labels = ['angry', 'sad']
-emodb.limit_samples = 150
 target = emotion
 [FEATS]
 type = ['audwav2vec2']
