+++ examples/exp_emodb_audmodel_xgb.ini
@@ -8,7 +8,7 @@ save = True
 databases = ['emodb']
 emodb = ./data/emodb/emodb
 emodb.split_strategy = random
-emodb.limit_samples = 200
+emodb.files_tables = ['emotion']
 emodb.mapping = {'anger':'angry', 'happiness':'happy', 'sadness':'sad', 'neutral':'neutral'}
 labels = ['angry', 'happy', 'neutral', 'sad']
 target = emotion
