+++ examples/exp_emodb_conformer.ini
@@ -12,7 +12,6 @@ emodb.test_tables = ['emotion.categories.test.gold_standard']
 emodb.train_tables = ['emotion.categories.train.gold_standard']
 labels = ['angry', 'happy', 'neutral', 'sad']
 emodb.mapping = {'anger':'angry', 'happiness':'happy', 'sadness':'sad', 'neutral':'neutral'}
-emodb.limit_samples = 350
 target = emotion
 [FEATS]
 #type = ['wav2vec2-conformer-rel-pos-large-960h-ft']
