+++ examples/exp_emodb_explore_data.ini
@@ -8,7 +8,7 @@ save = True
 databases = ['emodb']
 emodb = ./data/emodb/emodb
 emodb.split_strategy = random
-emodb.limit_samples = 100
+emodb.files_tables = ['emotion']
 target = emotion
 [FEATS]
 type = ['os']
