+++ examples/exp_emodb_explore_scatter.ini
@@ -9,7 +9,7 @@ no_warnings = True
 databases = ['emodb']
 emodb = ./data/emodb/emodb
 emodb.split_strategy = random
-emodb.limit_samples = 100
+emodb.files_tables = ['emotion']
 labels = ['angry', 'happy', 'neutral', 'sad']
 emodb.mapping = {'anger':'angry', 'happiness':'happy', 'sadness':'sad', 'neutral':'neutral'}
 target = emotion
