+++ examples/exp_emodb_os_praat_xgb.ini
@@ -7,8 +7,8 @@ save = True
 [DATA]
 databases = ['emodb']
 emodb = ./data/emodb/emodb
+emodb.files_tables = ['emotion']
 emodb.split_strategy = random
-emodb.limit_samples = 100
 emodb.mapping = {'anger':'angry', 'happiness':'happy', 'sadness':'sad', 'neutral':'neutral'}
 labels = ['angry', 'sad']
 target = emotion
