+++ examples/exp_emodb_predict.ini
@@ -14,7 +14,6 @@ labels = ['angry', 'happy', 'neutral', 'sad']
 #labels = ['disgust']
 emodb.mapping = {'anger':'angry', 'happiness':'happy', 'sadness':'sad', 'neutral':'neutral'}
 target = emotion
-emodb.limit_samples = 100
 [FEATS]
 type = ['os']
 store_format = csv
