+++ nkululeko/glob_conf.py
@@ -43,3 +43,11 @@ def set_labels(labels_obj):
 def set_target(target_obj):
     global target
     target = target_obj
+
+def set_split3(split3_obj):
+    global split3
+    split3 = split3_obj
+
+def set_got_speaker(got_speaker_obj):
+    global got_speaker
+    got_speaker = got_speaker_obj
